Use clean.js in library viewer
