+++ frontend/src/src.php
@@ -1,18 +1,13 @@
 <?php
-define('CLEANHOME', '/opt/clean');
+define('CLEAN_HOME', '/opt/clean');
 
-if (empty($_REQUEST['mod'])) :
+if (empty($_REQUEST['mod'])) {
 	echo '<p>Select a module on the left.</p>';
-else :
+	exit();
+}
 
 $mod = preg_replace('/[^\\w\\/\\. -`]/', '', $_REQUEST['mod']);
-
 $iclordcl = isset($_REQUEST['icl']) ? 'icl' : 'dcl';
-$hl_lines = isset($_REQUEST['line']) ? escapeshellarg($_REQUEST['line']) : '';
+$fname = CLEAN_HOME . '/lib/' . $mod . '.' . $iclordcl;
 
-$fname = CLEANHOME . '/lib/' . $mod . '.' . $iclordcl;
-$efname = escapeshellarg($fname);
-
-system("python3.5 cloogle_pygments.py $efname $hl_lines");
-
-endif;
+echo file_get_contents($fname);
